# Stack-based binary readers