--- a/src/AsmResolver.DotNet/Code/Cil/DynamicCilOperandResolver.cs
+++ b/src/AsmResolver.DotNet/Code/Cil/DynamicCilOperandResolver.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using AsmResolver.DotNet.Serialized;
 using AsmResolver.DotNet.Signatures;
+using AsmResolver.IO;
 using AsmResolver.PE.DotNet.Cil;
 using AsmResolver.PE.DotNet.Metadata.Tables;
 
@@ -83,9 +84,8 @@ namespace AsmResolver.DotNet.Code.Cil
                     break;
 
                 case TableIndex.StandAloneSig:
-                    return CallingConventionSignature.FromReader(
-                        new BlobReadContext(_readerContext),
-                        new ByteArrayReader((byte[]) _tokens[(int) token.Rid]));
+                    var reader = ByteArrayReaderFactory.CreateReader((byte[]) _tokens[(int) token.Rid]);
+                    return CallingConventionSignature.FromReader(new BlobReadContext(_readerContext), ref reader);
             }
 
             return token;