--- a/src/AsmResolver.DotNet/Serialized/SerializedFieldDefinition.cs
+++ b/src/AsmResolver.DotNet/Serialized/SerializedFieldDefinition.cs
@@ -12,7 +12,7 @@ namespace AsmResolver.DotNet.Serialized
 {
     /// <summary>
     /// Represents a lazily initialized implementation of <see cref="FieldDefinition"/>  that is read from a
-    /// .NET metadata image. 
+    /// .NET metadata image.
     /// </summary>
     public class SerializedFieldDefinition : FieldDefinition
     {
@@ -40,11 +40,20 @@ namespace AsmResolver.DotNet.Serialized
             .GetStringByIndex(_row.Name);
 
         /// <inheritdoc />
-        protected override FieldSignature GetSignature() => FieldSignature.FromReader(
-            new BlobReadContext(_context),
-            _context.Image.DotNetDirectory.Metadata
+        protected override FieldSignature GetSignature()
+        {
+            if (!_context.Image.DotNetDirectory.Metadata
                 .GetStream<BlobStream>()
-                .GetBlobReaderByIndex(_row.Signature));
+                .TryGetBlobReaderByIndex(_row.Signature, out var reader))
+            {
+                return _context.BadImageAndReturn<FieldSignature>(
+                    $"Invalid signature blob index in field {MetadataToken}.");
+            }
+
+            return FieldSignature.FromReader(
+                new BlobReadContext(_context),
+                ref reader);
+        }
 
         /// <inheritdoc />
         protected override TypeDefinition GetDeclaringType()
@@ -63,7 +72,7 @@ namespace AsmResolver.DotNet.Serialized
         /// <inheritdoc />
         protected override MarshalDescriptor GetMarshalDescriptor() =>
             _context.ParentModule.GetFieldMarshal(MetadataToken);
-        
+
         /// <inheritdoc />
         protected override ImplementationMap GetImplementationMap()
         {
@@ -78,13 +87,13 @@ namespace AsmResolver.DotNet.Serialized
         protected override ISegment GetFieldRva()
         {
             var module = _context.ParentModule;
-            
+
             uint rid = module.GetFieldRvaRid(MetadataToken);
             bool result = module.DotNetDirectory.Metadata
                 .GetStream<TablesStream>()
                 .GetTable<FieldRvaRow>()
                 .TryGetByRid(rid, out var fieldRvaRow);
-            
+
             if (!result)
                 return null;
 
@@ -100,7 +109,7 @@ namespace AsmResolver.DotNet.Serialized
                 .GetStream<TablesStream>()
                 .GetTable<FieldLayoutRow>()
                 .TryGetByRid(rid, out var fieldLayoutRow);
-            
+
             if (!result)
                 return null;
 
@@ -108,7 +117,7 @@ namespace AsmResolver.DotNet.Serialized
         }
 
         /// <inheritdoc />
-        protected override IList<CustomAttribute> GetCustomAttributes() => 
+        protected override IList<CustomAttribute> GetCustomAttributes() =>
             _context.ParentModule.GetCustomAttributeCollection(this);
     }
 }