… src/AsmResolver.DotNet/Serialized/SerializedMethodDefinition.cs
… src/AsmResolver.DotNet/Serialized/SerializedModuleDefinition.Metadata.cs
… src/AsmResolver.DotNet/Serialized/SerializedPropertyDefinition.cs
… src/AsmResolver.DotNet/Serialized/SerializedSecurityDeclaration.cs
… src/AsmResolver.DotNet/Serialized/SerializedStandAloneSignature.cs
… src/AsmResolver.DotNet/Serialized/SerializedTypeSpecification.cs
… src/AsmResolver.DotNet/Signatures/BlobSerializationContext.cs
… src/AsmResolver.DotNet/Signatures/CallingConventionSignature.cs
… src/AsmResolver.DotNet/Signatures/CustomAttributeArgument.cs
… src/AsmResolver.DotNet/Signatures/CustomAttributeArgumentReader.cs
… src/AsmResolver.DotNet/Signatures/CustomAttributeArgumentWriter.cs
… src/AsmResolver.DotNet/Signatures/CustomAttributeNamedArgument.cs
… src/AsmResolver.DotNet/Signatures/CustomAttributeSignature.cs
… src/AsmResolver.DotNet/Signatures/DataBlobSignature.cs
… src/AsmResolver.DotNet/Signatures/FieldSignature.cs
… src/AsmResolver.DotNet/Signatures/GenericInstanceMethodSignature.cs
… src/AsmResolver.DotNet/Signatures/LocalVariablesSignature.cs
… src/AsmResolver.DotNet/Signatures/Marshal/ComInterfaceMarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/Marshal/CustomMarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/Marshal/FixedArrayMarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/Marshal/FixedSysStringMarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/Marshal/LPArrayMarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/Marshal/MarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/Marshal/SafeArrayMarshalDescriptor.cs
… src/AsmResolver.DotNet/Signatures/MethodSignature.cs
… src/AsmResolver.DotNet/Signatures/MethodSignatureBase.cs
… src/AsmResolver.DotNet/Signatures/PropertySignature.cs
… src/AsmResolver.DotNet/Signatures/Security/PermissionSetSignature.cs
… src/AsmResolver.DotNet/Signatures/Security/SecurityAttribute.cs
… src/AsmResolver.DotNet/Signatures/Types/ArrayTypeSignature.cs