Closes #136