@Washi1337 2021-04-20
Initial testing reveals a pretty substantial increase in performance, in runtimes (around 22% faster) as well as memory consumption (up to at least 34% more efficient).

``` ini
BenchmarkDotNet=v0.12.1, OS=manjaro 
Intel Core i5-8300H CPU 2.30GHz (Coffee Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.202
  [Host]     : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
  DefaultJob : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
```

**Before:**

|                                Method |          Mean |       Error |        StdDev |        Median |     Gen 0 |     Gen 1 |     Gen 2 |   Allocated |
|-------------------------------------- |--------------:|------------:|--------------:|--------------:|----------:|----------:|----------:|------------:|
|                HelloWorld_PEFile_Read |      1.988 μs |   0.0218 μs |     0.0193 μs |      1.983 μs |    0.4692 |         - |         - |     1.93 KB |
|           HelloWorld_PEFile_ReadWrite |      5.590 μs |   0.0951 μs |     0.0843 μs |      5.555 μs |    1.7700 |         - |         - |     7.26 KB |
|               HelloWorld_PEImage_Read |      2.264 μs |   0.0398 μs |     0.0372 μs |      2.261 μs |    0.6638 |         - |         - |     2.71 KB |
|          HelloWorld_PEImage_ReadWrite |    200.603 μs |   2.1273 μs |     1.9899 μs |    200.843 μs |   35.1563 |    0.4883 |         - |   144.37 KB |
|      HelloWorld_ModuleDefinition_Read |    202.064 μs |   2.0577 μs |     1.7183 μs |    201.592 μs |   13.1836 |    0.2441 |         - |     54.4 KB |
| HelloWorld_ModuleDefinition_ReadWrite |  5,472.969 μs | 300.7044 μs |   867.6008 μs |  5,068.855 μs |         - |         - |         - |  6545.31 KB |
|                   CrackMe_PEFile_Read |      1.991 μs |   0.0347 μs |     0.0290 μs |      1.980 μs |    0.4692 |         - |         - |     1.93 KB |
|              CrackMe_PEFile_ReadWrite |      5.598 μs |   0.0653 μs |     0.0611 μs |      5.605 μs |    1.7700 |         - |         - |     7.26 KB |
|                  CrackMe_PEImage_Read |      2.291 μs |   0.0404 μs |     0.0378 μs |      2.273 μs |    0.6638 |         - |         - |     2.71 KB |
|             CrackMe_PEImage_ReadWrite |    205.676 μs |   2.4828 μs |     2.2010 μs |    205.054 μs |   35.4004 |    2.6855 |         - |   144.91 KB |
|         CrackMe_ModuleDefinition_Read |     51.132 μs |   0.6680 μs |     0.6249 μs |     51.098 μs |   13.0615 |    0.0610 |         - |    53.54 KB |
|    CrackMe_ModuleDefinition_ReadWrite | 17,119.715 μs | 351.1576 μs | 1,001.8724 μs | 16,917.233 μs | 1000.0000 | 1000.0000 | 1000.0000 | 26520.24 KB |


**After:**

|                                Method |          Mean |       Error |        StdDev |        Median |     Gen 0 |     Gen 1 |     Gen 2 |   Allocated |
|-------------------------------------- |--------------:|------------:|--------------:|--------------:|----------:|----------:|----------:|------------:|
|                HelloWorld_PEFile_Read |      1.988 μs |   0.0218 μs |     0.0193 μs |      1.983 μs |    0.4692 |         - |         - |     1.93 KB |
|           HelloWorld_PEFile_ReadWrite |      5.590 μs |   0.0951 μs |     0.0843 μs |      5.555 μs |    1.7700 |         - |         - |     7.26 KB |
|               HelloWorld_PEImage_Read |      2.264 μs |   0.0398 μs |     0.0372 μs |      2.261 μs |    0.6638 |         - |         - |     2.71 KB |
|          HelloWorld_PEImage_ReadWrite |    200.603 μs |   2.1273 μs |     1.9899 μs |    200.843 μs |   35.1563 |    0.4883 |         - |   144.37 KB |
|      HelloWorld_ModuleDefinition_Read |    202.064 μs |   2.0577 μs |     1.7183 μs |    201.592 μs |   13.1836 |    0.2441 |         - |     54.4 KB |
| HelloWorld_ModuleDefinition_ReadWrite |  5,472.969 μs | 300.7044 μs |   867.6008 μs |  5,068.855 μs |         - |         - |         - |  6545.31 KB |
|                   CrackMe_PEFile_Read |      1.991 μs |   0.0347 μs |     0.0290 μs |      1.980 μs |    0.4692 |         - |         - |     1.93 KB |
|              CrackMe_PEFile_ReadWrite |      5.598 μs |   0.0653 μs |     0.0611 μs |      5.605 μs |    1.7700 |         - |         - |     7.26 KB |
|                  CrackMe_PEImage_Read |      2.291 μs |   0.0404 μs |     0.0378 μs |      2.273 μs |    0.6638 |         - |         - |     2.71 KB |
|             CrackMe_PEImage_ReadWrite |    205.676 μs |   2.4828 μs |     2.2010 μs |    205.054 μs |   35.4004 |    2.6855 |         - |   144.91 KB |
|         CrackMe_ModuleDefinition_Read |     51.132 μs |   0.6680 μs |     0.6249 μs |     51.098 μs |   13.0615 |    0.0610 |         - |    53.54 KB |
|    CrackMe_ModuleDefinition_ReadWrite | 17,119.715 μs | 351.1576 μs | 1,001.8724 μs | 16,917.233 μs | 1000.0000 | 1000.0000 | 1000.0000 | 26520.24 KB |