@Washi1337 2021-04-21
Reading the metadata streams using the new struct reader seems to further reduce the memory footprint down to around 40%.

|                                Method |          Mean |       Error |        StdDev |        Median |     Gen 0 |     Gen 1 |     Gen 2 |   Allocated |
|-------------------------------------- |--------------:|------------:|--------------:|--------------:|----------:|----------:|----------:|------------:|
|                HelloWorld_PEFile_Read |      2.115 μs |   0.0073 μs |     0.0065 μs |      2.115 μs |    0.4730 |         - |         - |     1.94 KB |
|           HelloWorld_PEFile_ReadWrite |      6.181 μs |   0.1003 μs |     0.0889 μs |      6.152 μs |    3.7460 |         - |         - |    15.31 KB |
|               HelloWorld_PEImage_Read |      2.340 μs |   0.0061 μs |     0.0051 μs |      2.339 μs |    0.6638 |         - |         - |     2.72 KB |
|          HelloWorld_PEImage_ReadWrite |    193.676 μs |   3.8479 μs |     4.5807 μs |    193.395 μs |   37.1094 |    0.7324 |         - |   152.52 KB |
|      HelloWorld_ModuleDefinition_Read |    196.484 μs |   2.2836 μs |     2.1361 μs |    195.928 μs |   11.7188 |    0.2441 |         - |    48.57 KB |
| HelloWorld_ModuleDefinition_ReadWrite |  7,044.448 μs | 138.7240 μs |   115.8408 μs |  7,062.108 μs |         - |         - |         - |  5000.48 KB |
|                   CrackMe_PEFile_Read |      2.118 μs |   0.0155 μs |     0.0130 μs |      2.115 μs |    0.4730 |         - |         - |     1.94 KB |
|              CrackMe_PEFile_ReadWrite |      6.291 μs |   0.1245 μs |     0.2785 μs |      6.226 μs |    3.7460 |         - |         - |    15.31 KB |
|                  CrackMe_PEImage_Read |      2.411 μs |   0.0478 μs |     0.0447 μs |      2.417 μs |    0.6638 |         - |         - |     2.72 KB |
|             CrackMe_PEImage_ReadWrite |    194.487 μs |   2.2171 μs |     1.9654 μs |    194.635 μs |   37.3535 |    0.4883 |         - |   152.63 KB |
|         CrackMe_ModuleDefinition_Read |     50.153 μs |   0.7196 μs |     0.6731 μs |     50.330 μs |   11.5967 |    0.0610 |         - |    47.59 KB |
|    CrackMe_ModuleDefinition_ReadWrite | 12,971.299 μs | 522.2079 μs | 1,472.8967 μs | 12,458.904 μs | 2000.0000 | 1000.0000 | 1000.0000 | 15765.67 KB |