@Washi1337 2021-04-21
Ideally we would like to change the `CilRawMethodBody` implementation as well, and replace the `byte[] Code` property with a sliced reader struct as well, so we do not have to copy the raw CIL  bytes. However, this will be quite a significant breaking change on the public API, and for such a change the pros and cons have to be considered very carefully, which is probably out of scope for this PR.